Moved client code, added syntax highlighting to repl
